### 2018-09-14 #946 @whedon
[PRE REVIEW]: PyTeNet: A concise Python implementation of quantum tensor network algorithms
**Submitting author:** @cmendl (<a href="http://orcid.org/0000-0002-6386-0230">Christian B. Mendl</a>)
**Repository:** <a href="https://github.com/cmendl/pytenet" target ="_blank">https://github.com/cmendl/pytenet</a>
**Version:** v1.0
**Reviewers:** mlxd
