### 2018-09-14 #946 @cmendl
Concerning the article proof: would it be possible to switch to numerical citation style? I've added a reference to american-physics-society.csl in paper.md but apparently this wasn't used for compiling the paper to pdf...
