### 2018-09-14 #946 @jedbrown
@cmendl The citation style is meant to be standard across JOSS publications. I hope that's not a problem for you.

@jcmgray This is in a similar space to your recent article #819. Would you be available to review or could you suggest a colleague?

@mlxd We'd be thrilled if you could review this time.
