### 2018-09-15 #946 @mlxd
Sure, I can review this.
