### 2018-09-15 a1f6325
implement multiplication of two MPOs (composition along physical dimension)
