### 2018-09-15 #946 @jcmgray
I'd be more than happy to review, the only caveat being I am quite busy currently so it might be over the next few weeks.
