### 2018-09-15 #946 @jcmgray
@jedbrown
> add jcmgray **is** reviewer

I think this little typo might have missed me off as a reviewer.
