### 2018-09-15 53a6520
implement logical addition of two MPOs
