### 2018-09-17 7c371c3
MPO construction from operator chains now a classmethod
