+++ pytenet/mpo.py
@@ -18,44 +18,94 @@ class MPO(object):
     right virtual bond quantum number.
     """
 
-    def __init__(self, qd, **kwargs):
+    def __init__(self, qd, qD, fill=0.0):
         """
         Create a matrix product operator.
 
         Args:
             qd: physical quantum numbers at each site (same for all sites)
-        Keyword args: either provide oplist and L, or qD and (optionally) fill
-            oplist: list of operator chains
-            L: number of lattice sites (only accessed if oplist is provided)
             qD: virtual bond quantum numbers (list of quantum number lists)
-            fill: numerical value for filling the MPO tensors, or
-                 'random' for normally distributed random entries
+            fill: explicit scalar number to fill MPO tensors with, or
+                  'random' to initialize tensors with random complex entries
         """
-        # require NumPy array
+        # require NumPy arrays
         self.qd = np.array(qd)
+        self.qD = [np.array(qDi) for qDi in qD]
+        # create list of MPS tensors
         d = len(qd)
-        if 'opchains' in kwargs:
-            if not 'L' in kwargs:
-                raise ValueError('number of lattice sites L must be provided as keyword argument together with list of opchains')
-            self.from_opchains(d, kwargs['L'], kwargs['opchains'])
+        D = [len(qDi) for qDi in qD]
+        if isinstance(fill, int) or isinstance(fill, float) or isinstance(fill, complex):
+            self.A = [np.full((d, d, D[i], D[i+1]), fill) for i in range(len(D)-1)]
+        elif fill == 'random':
+            # random complex entries
+            self.A = [
+                    np.random.normal(size=(d, d, D[i], D[i+1]), scale=1./np.sqrt(d*D[i]*D[i+1])) +
+                 1j*np.random.normal(size=(d, d, D[i], D[i+1]), scale=1./np.sqrt(d*D[i]*D[i+1])) for i in range(len(D)-1)]
         else:
-            qD   = kwargs.get('qD',    [])
-            fill = kwargs.get('fill', 0.0)
-            D = [len(qb) for qb in qD]
-            self.qD = [np.array(qDi) for qDi in qD]
-            if isinstance(fill, int) or isinstance(fill, float) or isinstance(fill, complex):
-                self.A = [np.full((d, d, D[i], D[i+1]), fill) for i in range(len(D)-1)]
-            elif fill == 'random':
-                # random complex entries
-                self.A = [
-                        np.random.normal(size=(d, d, D[i], D[i+1]), scale=1./np.sqrt(d*D[i]*D[i+1])) +
-                     1j*np.random.normal(size=(d, d, D[i], D[i+1]), scale=1./np.sqrt(d*D[i]*D[i+1])) for i in range(len(D)-1)]
-            else:
-                raise ValueError('fill = {} invalid; must be a number or "random"'.format(fill))
-            # enforce block sparsity structure dictated by quantum numbers
-            for i in range(len(self.A)):
-                mask = qnumber_outer_sum([self.qd, -self.qd, self.qD[i], -self.qD[i+1]])
-                self.A[i] = np.where(mask == 0, self.A[i], 0)
+            raise ValueError('fill = {} invalid; must be a number or "random".'.format(fill))
+        # enforce block sparsity structure dictated by quantum numbers
+        for i in range(len(self.A)):
+            mask = qnumber_outer_sum([self.qd, -self.qd, self.qD[i], -self.qD[i+1]])
+            self.A[i] = np.where(mask == 0, self.A[i], 0)
+
+    @classmethod
+    def from_opchains(cls, qd, L, opchains):
+        """Construct a MPO representation of a sum of "operator chains"."""
+
[…]
