### 2018-09-23 cmendl/pytenet#1 @jcmgray
package pytenet
Looks like there is lot's of cool stuff here! Regarding https://github.com/openjournals/joss-reviews/issues/948:

It would be good if ``pytenet`` was structured as a standard install-able python package. Specifically, it needs a ``setup.py`` file with requirements etc. specified along with an ``__init__.py`` file in the main module subdirectory. Plenty of resources on this but there's a [very bare-bones example here](https://github.com/kennethreitz/samplemod).
