### 2018-09-25 #948 @mlxd
Hi  @cmendl
Really nice package. I have no problem recommending this for publicatoin, but I do have a few suggestions first based on the above checklist items:

**Version: Does the release version given match the GitHub release (v1.0)?**

No version is given for this software. Consider tagging the code with a release version for this     publication (eg `git tag v1.0`).

**Functionality
Installation: Does installation proceed as outlined in the documentation?**

No installation documentation provided for this package. As stated by @jmcgray above, it may be worth-while explaining how to install/run the software, using a package manager (conda/pip), or simply just cloning and running instructions.

**Functionality: Have the functional claims of the software been confirmed?**
The available physics of the software has been mentioned in the Readme.md file. However, giving a specific examples of each of the available functionalities may be worthwhile to anybody wishing to use the software. This could take the place of some Markdown file with example code sections, and some discussions of what is happening.

**Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.**

No installation instructions given. Please consider providing some installation instructions, either in the Readme.md, or elsewhere in documentation. See above comments for Functionality::Installation

**Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).**

Source examples provided in `pytenet/experiments` for several use-cases. Additionally, `pytenet/tests` directory has documented test-cases for many  routines. Consider adding a Readme.md to each respective directory to describe the listed files/classes/methods discussing each example usage.

**Functionality  documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?**

Source code has comments in key-locations to describe usage functionality. However, an API documentation framework may be beneficial (eg sphinx, doxygen, etc), to allow ease-of-use for developers and users. Alternatively, a Readme.md describing each source file/class/method in each respective directory would be beneficial.

**Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support**

No community guidelines given for development. Adding some options for options 1), 2), and 3) would be welcome.

Feel free to @ me for any further information if the above comments are insufficient.
