### 2018-09-25 #948 @cmendl
OK thanks for the suggestions! I'll try to incorporate them in the next few days..
