### 2018-09-25 #948 @jcmgray
I haven't yet had time to install and test the package (which looks like it has a lot of nice stuff!) but very much agree with all the documentation points @mlxd raised.

To just add a few thoughts, it would be very nice to have some 'getting started' style examples at a simpler level than the ``experiments`` but with more introduction and explanation than the tests. For example, a simple ground-state calculation, pure-state evolution and maybe a custom MPO hamiltonian construction might be good, minimal jumping off points.

I'm sure you know about jupyter notebooks, but just thought I'd mention that they can be a good choice for this kind of example (with possibly maths+code+graphs...) as github can render them ([here](https://github.com/jcmgray/quimb/blob/develop/docs/examples/ex_TEBD_evo.ipynb) is an example of mine) and they can even be incorporated into proper sphinx docs using [nbsphinx](https://nbsphinx.readthedocs.io).
