### 2018-09-25 cmendl/pytenet#1 @cmendl
Ok, thanks, I'll add that.

BTW I find the quimb toolbox very nice and elegant! (Admittedly didn't know about it before..)
