### 2018-09-25 cmendl/pytenet#1 @jcmgray
Nice, thanks.

> BTW I find the quimb toolbox very nice and elegant! (Admittedly didn't know about it before..)

Yeah it's hard to search through all these projects and I have not really been publicizing it! Of course if you see any opportunities for cross-pollination between the projects that would be cool - there are lots of features in ``pytenet`` that ``quimb`` does not yet implement.
