### 2018-09-28 416dc73
remodel PyTeNet into a Python package
