+++ pytenet/operation.py
@@ -1,5 +1,7 @@
 import numpy as np
 
+__all__ = ['operator_average', 'compute_right_operator_blocks', 'apply_local_hamiltonian', 'apply_local_bond_contraction']
+
 
 def operator_average(psi, op):
     """Compute the expectation value <psi | op | psi>.
