+++ test/test_operation.py
@@ -1,10 +1,6 @@
 import unittest
 import numpy as np
-import sys
-sys.path.append('../pytenet/')
-from mps import MPS
-from mpo import MPO
-from operation import operator_average
+import pytenet as ptn
 
 
 class TestOperation(unittest.TestCase):
@@ -19,7 +15,7 @@ class TestOperation(unittest.TestCase):
 
         # create random matrix product state
         D = [1, 7, 26, 19, 25, 8, 1]
-        psi = MPS(qd, [np.random.randint(-1, 2, size=Di) for Di in D], fill='random')
+        psi = ptn.MPS(qd, [np.random.randint(-1, 2, size=Di) for Di in D], fill='random')
         # rescale to achieve norm of order 1
         for i in range(psi.nsites):
             psi.A[i] *= 5
@@ -28,10 +24,10 @@ class TestOperation(unittest.TestCase):
         D = [1, 5, 16, 14, 17, 4, 1]
         # set bond quantum numbers to zero since otherwise,
         # sparsity pattern often leads to <psi | op | psi> = 0
-        op = MPO(qd, [np.zeros(Di) for Di in D], fill='random')
+        op = ptn.MPO(qd, [np.zeros(Di) for Di in D], fill='random')
 
         # calculate average (expectation value) <psi | op | psi>
-        avr = operator_average(psi, op)
+        avr = ptn.operator_average(psi, op)
 
         # reference value based on full Fock space representation
         x = psi.as_vector()
