### 2018-09-29 b30a825
rename eigh -> eigh_krylov and expm -> expm_krylov for clarity
