### 2018-09-30 cmendl/pytenet#1 @cmendl
resolved by commit 416dc73a1b6eae276d0139a82354aec6b0077406
