### 2018-10-06 3748810
revise and extend docstrings (Google Python style)
