### 2018-10-07 8bbca29
switch ReadMe from Markdown to reStructuredText format (for planned inclusion in API documentation)
