### 2018-10-07 10af59d
fix inclusion of ReadMe by setup.py
