### 2018-10-07 e1acdd3
files for Sphinx documentation
