+++ doc/requirements.txt
@@ -1,0 +1,2 @@
+sphinx>=1.8.1
+nbsphinx
