+++ doc/api.rst
@@ -1,0 +1,18 @@
+:orphan:
+
+API documentation
+=================
+
+.. autosummary::
+    :toctree: _autosummary
+
+    pytenet.bond_ops
+    pytenet.evolution
+    pytenet.hamiltonian
+    pytenet.krylov
+    pytenet.minimization
+    pytenet.mpo
+    pytenet.mps
+    pytenet.opchain
+    pytenet.operation
+    pytenet.qnumber
