### 2018-10-07 da6d78f
add link to online documentation in ReadMe; require ipython for syntax highlighting in .ipynb notebooks
