+++ doc/requirements.txt
@@ -1,2 +1,3 @@
 sphinx>=1.8.1
+ipython
 nbsphinx
