### 2018-10-08 8a72437
copy operator chains before padding to preserve original input
