### 2018-10-08 cf71edc
tutorial on quantum Hamiltonians as MPOs
