### 2018-10-08 #948 @cmendl
The raised issues should now be addressed: the package is now actually remodeled into a Python package for easy installation, I've included a short example code section for the TDVP time evolution directly in the ReadMe file, created two tutorials as Jupyter notebooks, and hosted them together with Sphinx API documentation at https://pytenet.readthedocs.io (@jcmgray quimb was really helpful as guide how to set this up...)

@mlxd : Before actually attaching the missing GitHub release (v1.0) tag, I'm planning to commit a slightly extended version of the manuscript. Should I preserve the original date or update it to the present date in paper.md? Thanks!
