### 2018-10-09 cmendl/pytenet#2 @jcmgray
make the basic example standalone?
[For review [here](https://github.com/openjournals/joss-reviews/issues/948)] All the new documentation is looking good! I just tried to run the basic example but it needs an MPS, ``psi``, to be defined. I tried what looked like the simplest matching constructor:
```python
psi = ptn.MPS(mpoH.qd, mpoH.qD, fill='random')
```
but I get a ``AssertionError`` during the lanczos iteration:
```python
...
assert beta[j] > 0
...
```
Zeroing the quantum numbers of the Hamiltonian, ``mpoH.zero_qnumbers()``, solves the issue however. My thoughts would be that it would be nice to have this snippet run without any tweaks as a starting point. Whether that means using the no-quantum numbers version or explaining how to set them up properly for the MPS (which I don't actually know yet!) is up to you.
