### 2018-10-10 7a3887a
gracefully ending Lanczos iteration (instead of failing) in case of beta[j] ~= 0
