### 2018-10-10 dd3661b
TDVP time evolution demo now self-contained (resolves #2)
