+++ ReadMe.rst
@@ -22,14 +22,27 @@ Example usage for TDVP time evolution (assuming `psi` is a matrix product state)
     # construct matrix product operator representation of
     # Heisenberg XXZ Hamiltonian (arguments are L, J, \Delta, h)
     mpoH = ptn.heisenberg_XXZ_MPO(L, 1.0, 0.8, -0.1)
+    mpoH.zero_qnumbers()
+    
+    # initial wavefunction as MPS with random entries
+    # maximally allowed virtual bond dimensions
+    D = [1, 2, 4, 8, 16, 28, 16, 8, 4, 2, 1]
+    psi = ptn.MPS(mpoH.qd, [Di*[0] for Di in D], fill='random')
+    # effectively clamp virtual bond dimension of initial state
+    Dinit = 8
+    for i in range(L):
+        psi.A[i][:, Dinit:, :] = 0
+        psi.A[i][:, :, Dinit:] = 0
+    psi.orthonormalize(mode='left')
+    
+    # time step can have both real and imaginary parts;
+    # for real time evolution use purely imaginary dt!
+    dt = 0.01 - 0.05j
+    numsteps = 100
     
     # run TDVP time evolution
-    # time step can have both real and imaginary parts;
-    # for real-time evolution use purely imaginary dt!
-    dt = 0.02 - 0.05j
-    numsteps = 100
     ptn.integrate_local_singlesite(mpoH, psi, dt, numsteps, numiter_lanczos=5)
-    # psi now stores the time-evolved state exp(-dt*numsteps H) psi
+    # psi now stores the (approximated) time-evolved state exp(-dt*numsteps H) psi
 
 
 Features
