### 2018-10-10 31c6353
minor revision of ReadMe file
