### 2018-10-10 #948 @mlxd
> @mlxd : Before actually attaching the missing GitHub release (v1.0) tag, I'm planning to commit a slightly extended version of the manuscript. Should I preserve the original date or update it to the present date in paper.md? Thanks!

If you are regenerating the paper with a longer version then I think an updated date should be fine (though Git always provides a paper trail of changes, so it probably shouldn't matter much). Otherwise, great work on the documentation and other changes.
