+++ paper/paper.bib
@@ -132,16 +132,6 @@
 
 @ARTICLE{WaltColbertVaroquaux2011,
   author = {van der Walt, S. and Colbert, S. C. and Varoquaux, G.},
-  title = {{The NumPy array: A structure for effcient numerical computation}},
-  journal = {Comput. Sci. Eng.},
-  year = {2011},
-  volume = {13},
-  pages = {22--30},
-  doi = {10.1109/MCSE.2011.37}
-}
-
-@ARTICLE{WaltColbertVaroquaux2011,
-  author = {van der Walt, S. and Colbert, S. C. and Varoquaux, G.},
   title = {{The NumPy array: A structure for efficient numerical computation}},
   journal = {Comput. Sci. Eng.},
   year = {2011},
@@ -169,3 +159,4 @@
   pages = {2863--2866},
   doi = {10.1103/PhysRevLett.69.2863}
 }
+
