### 2018-10-11 #948 @cmendl
Just updated the paper and added the tag...
