### 2018-10-16 #948 @mlxd
> ...so from my side, I've completed the current round of revision

No problem. I'll regenerate the PDF and we can go from there.
