### 2018-10-16 cmendl/pytenet#3 @mlxd
Add community guidelines
As part of the JOSS review process, the following check is required:

**Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Giving some indication on how to contribute to the code, some mention on best practices to add to the project, or to reach you can be added. This can be as simple as saying: "For new features please create an issue for discussion before creating a pull request", "Contact me at <email>, or create a GH issue", etc. so that anybody using your code can suggest features, file bugs, or the like.
