### 2018-10-16 #948 @mlxd
Guidelines issue created: https://github.com/cmendl/pytenet/issues/3
