### 2018-10-16 71a92b9
add "Contributing" section to ReadMe (resolves #3)
