### 2018-10-16 #948 @mlxd
I'm happy with the current version. Once @jcmgray is too I think we can end the review.
