### 2018-10-16 #948 @jcmgray
Yes I've had a look through and very happy with all the changes - so good to go on my part!
